Because, unsurprisingly the 'shuffle' to solve option didn't work, I had to go another route.
I started by implementing something called the 'homogeny' index on each side. This index is calculated by seeing how many same colors are located on a side at any given time.

The next piece of this will be to observe the results of a movement as some aggregate of this number, and then decide from that, which is the best move to do.